+++ lib/post/pages/post_page.dart
@@ -394,6 +394,7 @@ class _PostPageState extends State<PostPage> {
                                 comments: state.comments,
                                 selectedCommentId: state.selectedCommentId,
                                 selectedCommentPath: state.selectedCommentPath,
+                                newlyCreatedCommentId: state.newlyCreatedCommentId,
                                 moddingCommentId: state.moddingCommentId,
                                 viewFullCommentsRefreshing: state.viewAllCommentsRefresh,
                                 itemScrollController: _itemScrollController,
@@ -492,6 +493,9 @@ class _PostPageState extends State<PostPage> {
     AuthBloc authBloc = context.read<AuthBloc>();
     AccountBloc accountBloc = context.read<AccountBloc>();
 
+    final ThunderState state = context.read<ThunderBloc>().state;
+    final bool reduceAnimations = state.reduceAnimations;
+
     if (!authBloc.state.isLoggedIn) {
       showSnackbar(context, AppLocalizations.of(context)!.mustBeLoggedInComment);
     } else {
@@ -512,6 +516,7 @@ class _PostPageState extends State<PostPage> {
       Navigator.of(context)
           .push(
         SwipeablePageRoute(
+          transitionDuration: reduceAnimations ? const Duration(milliseconds: 100) : null,
           canOnlySwipeFromEdge: true,
           backGestureDetectionWidth: 45,
           builder: (context) {