+++ lib/post/utils/comment_actions.dart
@@ -49,6 +49,9 @@ void triggerCommentAction({
       ThunderBloc thunderBloc = context.read<ThunderBloc>();
       AccountBloc accountBloc = context.read<AccountBloc>();
 
+      final ThunderState state = context.read<ThunderBloc>().state;
+      final bool reduceAnimations = state.reduceAnimations;
+
       SharedPreferences prefs = (await UserPreferences.instance).sharedPreferences;
       DraftComment? newDraftComment;
       DraftComment? previousDraftComment;
@@ -66,6 +69,7 @@ void triggerCommentAction({
       Navigator.of(context)
           .push(
         SwipeablePageRoute(
+          transitionDuration: reduceAnimations ? const Duration(milliseconds: 100) : null,
           canOnlySwipeFromEdge: true,
           backGestureDetectionWidth: 45,
           builder: (context) {