+++ lib/post/widgets/comment_view.dart
@@ -25,6 +25,7 @@ class CommentSubview extends StatefulWidget {
   final PostViewMedia? postViewMedia;
   final int? selectedCommentId;
   final String? selectedCommentPath;
+  final int? newlyCreatedCommentId;
   final int? moddingCommentId;
   final ItemScrollController itemScrollController;
   final ItemPositionsListener itemPositionsListener;
@@ -46,6 +47,7 @@ class CommentSubview extends StatefulWidget {
     this.postViewMedia,
     this.selectedCommentId,
     this.selectedCommentPath,
+    this.newlyCreatedCommentId,
     this.moddingCommentId,
     required this.itemScrollController,
     required this.itemPositionsListener,
@@ -94,6 +96,8 @@ class _CommentSubviewState extends State<CommentSubview> with SingleTickerProvid
     final theme = Theme.of(context);
     final ThunderState state = context.read<ThunderBloc>().state;
 
+    final reduceAnimations = state.reduceAnimations;
+
     if (!widget.viewFullCommentsRefreshing && _removeViewFullCommentsButton) {
       _animatingIn = true;
       _fullCommentsAnimation.reverse();
@@ -107,9 +111,17 @@ class _CommentSubviewState extends State<CommentSubview> with SingleTickerProvid
             duration: const Duration(milliseconds: 250),
             curve: Curves.easeInOut,
           );
+        } else if (state.newlyCreatedCommentId != null && state.comments.first.commentView?.comment.id == state.newlyCreatedCommentId) {
+          // Only scroll for top level comments since you can comment from anywhere in the comment section.
+          widget.itemScrollController.scrollTo(
+            index: 1,
+            duration: const Duration(milliseconds: 250),
+            curve: Curves.easeInOut,
+          );
         }
       },
       child: ScrollablePositionedList.builder(
+        physics: reduceAnimations ? const BouncingScrollPhysics() : null,
         addSemanticIndexes: false,
         itemScrollController: widget.itemScrollController,
         itemPositionsListener: widget.itemPositionsListener,
@@ -172,6 +184,7 @@ class _CommentSubviewState extends State<CommentSubview> with SingleTickerProvid
                       now: widget.now,
                       selectCommentId: widget.selectedCommentId,
                       selectedCommentPath: widget.selectedCommentPath,
+                      newlyCreatedCommentId: widget.newlyCreatedCommentId,
                       moddingCommentId: widget.moddingCommentId,
                       commentViewTree: widget.comments[index - 1],
                       collapsedCommentSet: collapsedCommentSet,