+++ lib/routes.dart
@@ -7,6 +7,7 @@ import 'package:thunder/account/bloc/account_bloc.dart';
 import 'package:thunder/core/auth/bloc/auth_bloc.dart';
 
 import 'package:thunder/settings/pages/about_settings_page.dart';
+import 'package:thunder/settings/pages/accessibility_settings_page.dart';
 import 'package:thunder/settings/pages/debug_settings_page.dart';
 import 'package:thunder/settings/pages/fab_settings_page.dart';
 import 'package:thunder/settings/pages/general_settings_page.dart';
@@ -67,6 +68,16 @@ final GoRouter router = GoRouter(
             return BlocProvider.value(
               value: state.extra! as ThunderBloc,
               child: const FabSettingsPage(),
+            );
+          },
+        ),
+        GoRoute(
+          name: 'accessibility',
+          path: 'accessibility',
+          builder: (context, state) {
+            return BlocProvider.value(
+              value: state.extra! as ThunderBloc,
+              child: const AccessibilitySettingsPage(),
             );
           },
         ),